HSEARCH-2387 Move Elasticsearch to an SPI package
