+++ elasticsearch/src/main/java/org/hibernate/search/elasticsearch/bridge/spi/Elasticsearch.java
@@ -4,10 +4,11 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.elasticsearch.bridge.builtin.impl;
+package org.hibernate.search.elasticsearch.bridge.spi;
 
 import org.hibernate.search.bridge.spi.FieldMetadataCreationContext;
 import org.hibernate.search.bridge.spi.FieldType;
+import org.hibernate.search.elasticsearch.bridge.builtin.impl.ElasticsearchBridgeDefinedField;
 import org.hibernate.search.elasticsearch.cfg.DynamicType;
 import org.hibernate.search.engine.metadata.impl.BridgeDefinedField;
